+++ src/app/collection/textArea/TextMain.jsx
@@ -1,30 +1,30 @@
-import React, {useEffect, useMemo} from 'react';
+import React, {useEffect, useLayoutEffect, useMemo, useState} from 'react';
 import styled from "styled-components";
-import {useDispatch} from "react-redux";
-import {addAddAnimation,} from "../../../redux/slice/collectionSlice";
+import {useDispatch, useSelector} from "react-redux";
+import {addAddAnimation, addOpenEditToolbar, selectCollection,} from "../../../redux/slice/collectionSlice";
 import {global, autoFocus} from "../../../assets/global";
+import Comment from "../../../component/note/Comment.jsx";
+import Textarea from "../../../component/note/Textarea.jsx";
+import Url from "../../../component/note/Url.jsx";
+import Note from '../../../component/note/Note.jsx'
+import ThisIsBottom from "../../../component/ThisIsBottom.jsx";
+import Warning from "../../../component/Warning.jsx";
+import {DragDropContext, Draggable, Droppable} from "react-beautiful-dnd";
 
-import Toolbar from "../../component/Toolbar.jsx";
-import Comment from "../../component/Comment.jsx";
-import TextareaOfTextarea from "../../component/TextareaOfTextarea.jsx";
-import Url from "../../component/Url.jsx";
-import Note from '../../component/Note.jsx'
-import ThisIsBottom from "../../component/ThisIsBottom.jsx";
-import Warning from "../../component/Warning.jsx";
 
-const {tertiary, secondary} = global
+const {main, tertiary, secondary, quaternary, max_height, max_width, font_size_m} = global
 
 
 /* main */
 const TextMain = styled.div`
-    width: 100%;
-    height: calc(100% - 48px);
+    position: fixed;
+    width: ${max_width}px;
+    height: ${max_height}px;
     display: flex;
     flex-direction: column;
     align-items: center;
     padding: 5px 10px 10px 10px;
-    overflow: scroll;
-    overflow-X: hidden;
+    overflow: scroll; 
     &::-webkit-scrollbar {
         width: 5px;
     }
@@ -37,20 +37,13 @@ const TextMain = styled.div`
 
 /* 底部 */
 
-
-
+const aa = ['aa', 'bb', 'cc']
 export default function App(p) {
     const dispatch = useDispatch()
     const {saveWarning, id, obj} = p
+    const {addNewNoteAnimation, addNewCommentAnimation, openEditId} = useSelector(selectCollection)
     const data = obj[id]
-    const {addNewNoteAnimation, openEditId} = obj
-
-    /* textarea 根據內容展開 */
-    // useEffect(() => {
-    //     document.querySelectorAll('textarea').forEach(item => {
-    //         item.style.height = item.scrollHeight + 'px'
-    //     })
-    // }, [data]);
+    const [isOpenComment, setIsOpenComment] = useState(false)
 
 
     /* 新增筆記後的動畫 */
@@ -60,40 +53,88 @@ export default function App(p) {
         }, 0)
     }, [addNewNoteAnimation])
 
+
+    // return (
+    //     <TextMain>
+    //         <DragDropContext>
+    //             <Droppable
+    //                 key={'item.key'}
+    //                 droppableId={`drog_key_s`}
+    //                 // isDropDisabled={!(openEditId === item.key)}
+    //             >
+    //                 {(provided) => (
+    //                     <div
+    //                         ref={provided.innerRef}
+    //                         {...provided.droppableProps}
+    //                     >
+    //                         {aa.map((item, idx) => (
+    //                             <Draggable
+    //                                 key={`drag_key_${idx}`}
+    //                                 draggableId={`draggableId_${idx}`}
+    //                                 index={idx}
+    //                             >
+    //                                 {
+    //                                     (provided) => (
+    //                                         <div
+    //                                             ref={provided.innerRef}
+    //                                             {...provided.draggableProps}
+    //                                             {...provided.dragHandleProps}
+    //                                         >
+    //                                             <div>{item}</div>
+    //
+    //                                         </div>
+    //                                     )
+    //                                 }
+    //                             </Draggable>
+    //                         ))}
+    //                         {provided.placeholder}
+    //                     </div>
+    //                 )}
+    //             </Droppable>
+    //         </DragDropContext>
+    //     </TextMain>
+    // );
+
+
     return (
         <TextMain>
-            <Warning
-                warning={saveWarning}>
-                請退出編輯後再離開
-            </Warning>
-            {
-                data.map(item => (
-                    <Note
-                        key={item.key}
-                        item={item}
-                        // Toolbar={
-                        //     <Toolbar
-                        //         item={item}
-                        //         area={'collection'}
-                        //         noteType={item.type === 'collection'}/>
-                        // }
-                        TextareaOfTextarea={
-                            <TextareaOfTextarea
-                                item={item}/>
-                        }
-                        Url={
-                            <Url
-                                item={item}/>
-                        }
-                        Comment={
-                            item.comment
-                            && <Comment
-                                comment={item.comment}
-                                obj={obj}/>
-                        }
-                    />
-                ))
-            }
+            <Warning warning={saveWarning}>請退出編輯後再離開</Warning>
+            <DragDropContext>
+                {
+                    data.map(item => (
+                        <Droppable
+                            key={item.key}
+                            droppableId={`drog_key_${item.key}`}
+                            isDropDisabled={!(openEditId === item.key)}
+                        >
+                            {(provided) => (
+                                <div
+                                    ref={provided.innerRef}
+                                    {...provided.droppableProps}
+                                >
+                                    <Note
+                                        item={item}
+                                        isOpenComment={isOpenComment}
+                                    >
+                                        <Textarea item={item}/>
+                                        <Url item={item}/>
+                                        {
+                                            (item.comment.length > 0)
+                                            && <Comment
+                                                parentItem={item}
+                                                comment={item.comment}
+                                                obj={obj}
+                                                dragHandle={{...provided.dragHandleProps}}
+                                            />
+                                        }
+                                    </Note>
+                                    {provided.placeholder}
+                                </div>
+                            )}
+                        </Droppable>
+                    ))
+                }
+            </DragDropContext>
             <ThisIsBottom/>
         </TextMain>
     );