+++ src/app/folder/FolderName.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from "styled-components";
 import {global} from "../../assets/global";
 
-const {primary, tertiary} = global
+const {primary, tertiary, font_size_m} = global
 const Page = styled.div`
     text-align: center;
     color: ${({font}) => font ? 'black' : 'white'};
@@ -11,7 +11,7 @@ const Page = styled.div`
     border: none;
     margin: 10px 0;
     overflow: hidden;
-    font-size: 12px;
+    font-size: ${font_size_m}px;
 `
 function FolderName(props) {
     const {font} = props