+++ src/assets/fakeData.js
@@ -75,13 +75,66 @@ const N1 = [
     //     ]
     // },
     {
-        key: '1556556',
+        key: 'Juuten_999556',
+        type: 'collection',
+        favIconUrl: 'https://img.icons8.com/material/344/globe--v2.png',
+        url: 'https://www.youtube.com/watch?v=HSW6qfRUT9w',
+        currentDate: '2022/10/10 12:11',
+        msg: `{"blocks":[{"key":"an8qb","text":"","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}}],"entityMap":{}}`,
+        comment: [
+            {
+                key: 'Juuten_255',
+                type: 'comment',
+                currentDate: '2022/10/10 12:11',
+                msg: '{"blocks":[{"key":"70fru","text":"最慘的難，去了已經看不到俄羅斯的未來在哪裡...","type":"unstyled","depth":0,"inlineStyleRanges":[{"offset":44,"length":3,"style":"CODE"},{"offset":106,"length":3,"style":"CODE"}],"entityRanges":[],"data":{}}],"entityMap":{}}',
+                comment: []
+            },
+            {
+                key: 'Juuten_344',
+                type: 'comment',
+                currentDate: '2022/10/10 12:11',
+                msg: '{"blocks":[{"key":"1ebtn","text":"對絕大部分整個烏克嚴重悟（或許吧）","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}}],"entityMap":{}}',
+                comment: []
+            },
+        ]
+    },
+    {
+        key: 'Juuten_1556556',
         type: 'collection',
         favIconUrl: 'https://img.icons8.com/material/344/globe--v2.png',
         url: 'https://www.youtube.com/watch?v=HSW6qfRUT9w',
         currentDate: '2022/10/10 12:11',
         msg: `{"blocks":[{"key":"8s8rf","text":"本影片的資料主要來源為網上與社群媒體，資料眾","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}},{"key":"fuqhl","text":"多未能一一查","type":"header-three","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}},{"key":"ctr4","text":"，資訊並不一定百分之百正確，內容僅供參考。志在分享，如有誤點，歡迎指正。","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}}],"entityMap":{}}`,
-        comment: []
+        comment: [
+            {
+                key: 'Juuten_2',
+                type: 'comment',
+                currentDate: '2022/10/10 12:11',
+                msg: '{"blocks":[{"key":"70fru","text":"最慘的是永遠是俄羅斯自己蘇聯時期跑掉一堆人才，導致解體後除了能源軍工外各個產業發展困難，GDP窮到只跟伊朗差不多結果開戰後變本加厲，不只跑掉更多人才，連唯二賺錢的軍火和能源都賣不出去了已經看不到俄羅斯的未來在哪裡...","type":"unstyled","depth":0,"inlineStyleRanges":[{"offset":44,"length":3,"style":"CODE"},{"offset":106,"length":3,"style":"CODE"}],"entityRanges":[],"data":{}}],"entityMap":{}}',
+                comment: []
+            },
+            {
+                key: 'Juuten_3',
+                type: 'comment',
+                currentDate: '2022/10/10 12:11',
+                msg: '{"blocks":[{"key":"1ebtn","text":"對絕大部分平民來說，哪個地方是不是自己的，與個人利益幾乎完全沒有任何關係，就假如俄羅斯真的奪下整個烏克蘭，中國奪下整個台灣，國內的人民的生活絲毫不會有任何改善，甚至可能承受種種戰爭下的苦果，但對於那些已被嚴重洗腦的人來說講再多也是白費唇舌，只有真正體會了才能幡然醒悟（或許吧）","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}}],"entityMap":{}}',
+                comment: []
+            },
+            {
+                key: 'Juuten_5',
+                type: 'comment',
+                currentDate: '2022/10/10 12:11',
+                msg: '{"blocks":[{"key":"3spv9","text":"舉例有些遙遠，個人是看長期國際貿易和貿易聯盟零關稅加深供應鏈過度集中，供應堅韌性逐漸下降一堆國家工業製造都出現斷層，各類產品可能主要前3產國就可以占到市占7~9成，如同投資放在同個籃利潤大風險也大，現在風險發生利潤是要還回去的可能還要倒還，就如同長期政客為了當選低利率，風險發生通膨要全民一起還，俄國出口主要的自然資源又是附近國家很缺的+加上國貿較晚起步還不夠依賴，影響自然會比較少；現在中台貿易從現政府上任一來一年3.5兆漲到7兆台幣都不知道到底是不是假站美國邊，然後台幣升就阻時不時升跌就放跌...真的超獨厚出口資方老闆；供應堅韌性逐漸下降EX:疫情如果在晚個5年爆發搞不好那時台灣國內就已經沒有人生產口罩了","type":"unstyled","depth":0,"inlineStyleRanges":[{"offset":67,"length":1,"style":"CODE"},{"offset":77,"length":3,"style":"CODE"},{"offset":168,"length":1,"style":"CODE"},{"offset":208,"length":3,"style":"CODE"},{"offset":214,"length":1,"style":"CODE"},{"offset":248,"length":3,"style":"CODE"},{"offset":272,"length":3,"style":"CODE"},{"offset":282,"length":1,"style":"CODE"}],"entityRanges":[],"data":{}}],"entityMap":{}}',
+                comment: []
+            },
+            //         {
+            //             key: 4,
+            //             type: 'comment',
+            //             currentDate: '2022/10/10 12:11',
+            //             msg: '如果沒有引發第三次世界大戰，我可以預測，下一次的諾貝爾文學獎，會由烏克蘭人獲得。',
+            //             comment: []
+            //         }
+        ]
     },
 ]
 
@@ -154,5 +207,39 @@ const N1 = [
 //         comment: []
 //     }
 // ]
-const Juuten_Storage = []
+const Juuten_Storage = [
+    {
+        key: 'Juuten_23',
+        msg: '{"blocks":[{"key":"327mv","text":"hi i am collection","type":"unstyled","depth":0,"inlineStyleRanges":[],"entityRanges":[],"data":{}}],"entityMap":{}}',
+        type: 'storage',
+        url: 'https://google.com',
+        favIconUrl: '',
+        pageTitle: 'david love lucy',
+        currentDate: '2022/10/10 12:10',
+        position: 203,
+        comment: []
+    },
+    {
+        key: 'Juuten_24',
+        msg: '{"blocks":[{"key":"8ot9e","text":"巴哈姆特 GNN 與日本遊戲業界分析公司 MediaCreate 合作，每週定期刊載由 MediaCreate 調查與統計的日本、台灣與韓國地區家用遊樂器主機遊戲軟體的一週銷售排行榜，供玩家參考","type":"unstyled","depth":0,"inlineStyleRanges":[{"offset":4,"length":5,"style":"CODE"},{"offset":20,"length":13,"style":"CODE"},{"offset":43,"length":13,"style":"CODE"}],"entityRanges":[],"data":{}}],"entityMap":{}}',
+        type: 'storage',
+        url: 'https://gnn.gamer.com.tw/detail.php?sn=240417',
+        favIconUrl: '',
+        pageTitle: 'MediaCreate 10 月 17 日～10 月 23 日一週銷售排行榜 《女神異聞錄 5 皇家版》包辦冠亞季軍',
+        currentDate: '2022/10/10 12:10',
+        position: 203,
+        comment: []
+    },
+//     {
+//         key: 25,
+//         msg: '弗利沙宇宙船坐墊（暫譯，ドラゴンボールZ ダメになる!? フリーザポッドクッション）」的尺寸寬約為 95 公分、高 70 公分，可容納一名成人',
+//         type: 'collection',
+//         url: 'https://gnn.gamer.com.tw/detail.php?sn=240381',
+//         favIconUrl: '',
+//         pageTitle: 'david love lucy',
+//         currentDate: '2022/10/10 12:10',
+//         position: 203,
+//         comment: []
+//     },
+]
 export {Juuten_Storage, N1, Juuten_folderLists}