+++ src/assets/global.js
@@ -14,11 +14,17 @@ export const global = {
     tertiary: '#8A8A8A',
     quaternary: '#eff3f4',
     warning: '#9d201a',
+    max_width: 500,
+    max_height: 600,
     transition_speed1: 'transition: 0.2s ease-out;',
+    transition_speed1_delay: 'transition: 0.5s ease-out 1s;',
     transition_speed2: 'transition: 0.5s ease-out;',
     icon_size_s: 12,
     icon_size_m: 15,
     icon_size_l: 18,
+    font_size_s: 10,
+    font_size_m: 12,
+    font_size_l: 15,
 
 
     color: [